+++ assets/app-config.js
@@ -1,2 +1,2 @@
-var ASSET_VER = '7475';
+var ASSET_VER = '7476';
 window.ASSET_VER = ASSET_VER;